* master: (43 commits)
  Syscall specific hooks (#2389)
  TUI Support Infrastructure (#1620)
  Fix coveralls upload (#2387)
  docs: fix simple typo, straigth -> straight (#2381)
  Attempt to allow symbolic balances from the start (#1818)
  Fix state.cpu.PC member (#1825)
  Bump black and mypy (#1824)
  Manticore 0.3.5 (#1808)
  Fix yices timeout argument (#1817)
  Detect default solver (#1820)
  Ignore Gas Calculations by Default (#1816)
  native/cpu/x86: Add support for CPUID EAX=80000000h (#1811)
  Change types.FunctionType=<class 'function'> (#1803)
  Fix test regressions (#1804)
  State Introspection API (#1775)
  Fix EVM account existence checks for selfdestruct and call (#1801)
  Add partial implementation of sendto syscall (#1791)
  crytic-compile: use latest release (#1795)
  Update gas metering for calls to empty accounts (#1774)
  Fix BitVec with symbolic offset and fix TranslatorSmtlib.unique thread safety (#1792)
  ...

+++ tests/wasm_sym/generate_symbolic_tests.sh
@@ -3,15 +3,19 @@
 rm -rf */
 touch __init__.py
 
-wget -nc -nv -O wabt.tgz -c https://github.com/WebAssembly/wabt/releases/download/1.0.12/wabt-1.0.12-linux.tar.gz
+if ! [ -x "$(command -v wast2json)" ]; then
+  wget -nc -nv -O wabt.tgz -c https://github.com/WebAssembly/wabt/releases/download/1.0.12/wabt-1.0.12-linux.tar.gz
+  tar --wildcards --strip=1 -xf wabt.tgz 'wabt-*/wast2json'
+  rm wabt.tgz
+else
+  cp "$(command -v wast2json)" .
+fi
+
 wget -nc -nv -O spec.zip -c https://github.com/WebAssembly/spec/archive/opam-1.1.zip
 
-yes | unzip -q -j spec.zip 'spec-*/test/core/*' -d .
+unzip -q -j spec.zip 'spec-*/test/core/*' -d .
 rm run.py README.md
 rm spec.zip
-
-tar --wildcards --strip=1 -xf wabt.tgz 'wabt-*/wast2json'
-rm wabt.tgz
 
 mkdir skipped_tests
 while read skip; do
